восстанавливаю пустой поинт
